+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Improve check for expired certs in the sp_expiration collector.
+
 ## [2.10.0] - 2022-03-21
 
 ### Added